third_party/WebKit/LayoutTests/external/wpt/cookies/http-state/0004-expected.txt

This is a testharness.js-based test.
FAIL Set cookie without value. assert_equals: expected "" but got "foo"
Harness: the test ran to completion.
